+++ src/test/java/it/pagopa/pn/paperchannel/rest/v1/PaperChannelRestV1ControllerTest.java
@@ -11,6 +11,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import reactor.core.publisher.Mono;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -251,8 +252,8 @@ class PaperChannelRestV1ControllerTest {
         request.setUid("UUID");
         request.setTenderCode("tenderCode");
         request.setDriverCode("driverCode");
-        request.setPrice(10.3f);
-        request.setPriceAdditional(11.2f);
+        request.setPrice(BigDecimal.valueOf(10.3f));
+        request.setPriceAdditional(BigDecimal.valueOf(11.2f));
         request.setProductType(ProductTypeEnumDto.AR);
         request.setCap(caps);
         request.setZone(InternationalZoneEnum._1);
